requirements.txt
pandas==1.2.4
numpy==1.20.1
tweepy==4.4.0
python-binance==1.0.16
finplot==1.8.1
